# Fix arrow perf regression