--- a/cpp/src/io/parquet/ipc/Schema_generated.h
+++ b/cpp/src/io/parquet/ipc/Schema_generated.h
@@ -5,11 +5,11 @@
 
 #include <flatbuffers/flatbuffers.h>
 
-// Ensure the included flatbuffers.h is the same version as when this file was
-// generated, otherwise it may not be compatible.
-static_assert(FLATBUFFERS_VERSION_MAJOR == 24 && FLATBUFFERS_VERSION_MINOR == 3 &&
-                FLATBUFFERS_VERSION_REVISION == 25,
-              "Non-compatible flatbuffers version included");
+//// Ensure the included flatbuffers.h is the same version as when this file was
+//// generated, otherwise it may not be compatible.
+//static_assert(FLATBUFFERS_VERSION_MAJOR == 24 && FLATBUFFERS_VERSION_MINOR == 3 &&
+//                FLATBUFFERS_VERSION_REVISION == 25,
+//              "Non-compatible flatbuffers version included");
 
 namespace cudf {
 namespace io {