--- a/java/src/main/native/src/ColumnVectorJni.cpp
+++ b/java/src/main/native/src/ColumnVectorJni.cpp
@@ -37,69 +37,11 @@
 #include <cudf/utilities/memory_resource.hpp>
 
 #include <arrow/api.h>
-#include <arrow/c/bridge.h>
 
 #include <algorithm>
 
 using cudf::jni::ptr_as_jlong;
 using cudf::jni::release_as_jlong;
-
-// Creating arrow as per given type_id and buffer arguments
-template <typename... Ts>
-std::shared_ptr<arrow::Array> to_arrow_array(cudf::type_id id, Ts&&... args)
-{
-  switch (id) {
-    case cudf::type_id::BOOL8:
-      return std::make_shared<arrow::BooleanArray>(std::forward<Ts>(args)...);
-    case cudf::type_id::INT8: return std::make_shared<arrow::Int8Array>(std::forward<Ts>(args)...);
-    case cudf::type_id::INT16:
-      return std::make_shared<arrow::Int16Array>(std::forward<Ts>(args)...);
-    case cudf::type_id::INT32:
-      return std::make_shared<arrow::Int32Array>(std::forward<Ts>(args)...);
-    case cudf::type_id::INT64:
-      return std::make_shared<arrow::Int64Array>(std::forward<Ts>(args)...);
-    case cudf::type_id::UINT8:
-      return std::make_shared<arrow::UInt8Array>(std::forward<Ts>(args)...);
-    case cudf::type_id::UINT16:
-      return std::make_shared<arrow::UInt16Array>(std::forward<Ts>(args)...);
-    case cudf::type_id::UINT32:
-      return std::make_shared<arrow::UInt32Array>(std::forward<Ts>(args)...);
-    case cudf::type_id::UINT64:
-      return std::make_shared<arrow::UInt64Array>(std::forward<Ts>(args)...);
-    case cudf::type_id::FLOAT32:
-      return std::make_shared<arrow::FloatArray>(std::forward<Ts>(args)...);
-    case cudf::type_id::FLOAT64:
-      return std::make_shared<arrow::DoubleArray>(std::forward<Ts>(args)...);
-    case cudf::type_id::TIMESTAMP_DAYS:
-      return std::make_shared<arrow::Date32Array>(std::make_shared<arrow::Date32Type>(),
-                                                  std::forward<Ts>(args)...);
-    case cudf::type_id::TIMESTAMP_SECONDS:
-      return std::make_shared<arrow::TimestampArray>(arrow::timestamp(arrow::TimeUnit::SECOND),
-                                                     std::forward<Ts>(args)...);
-    case cudf::type_id::TIMESTAMP_MILLISECONDS:
-      return std::make_shared<arrow::TimestampArray>(arrow::timestamp(arrow::TimeUnit::MILLI),
-                                                     std::forward<Ts>(args)...);
-    case cudf::type_id::TIMESTAMP_MICROSECONDS:
-      return std::make_shared<arrow::TimestampArray>(arrow::timestamp(arrow::TimeUnit::MICRO),
-                                                     std::forward<Ts>(args)...);
-    case cudf::type_id::TIMESTAMP_NANOSECONDS:
-      return std::make_shared<arrow::TimestampArray>(arrow::timestamp(arrow::TimeUnit::NANO),
-                                                     std::forward<Ts>(args)...);
-    case cudf::type_id::DURATION_SECONDS:
-      return std::make_shared<arrow::DurationArray>(arrow::duration(arrow::TimeUnit::SECOND),
-                                                    std::forward<Ts>(args)...);
-    case cudf::type_id::DURATION_MILLISECONDS:
-      return std::make_shared<arrow::DurationArray>(arrow::duration(arrow::TimeUnit::MILLI),
-                                                    std::forward<Ts>(args)...);
-    case cudf::type_id::DURATION_MICROSECONDS:
-      return std::make_shared<arrow::DurationArray>(arrow::duration(arrow::TimeUnit::MICRO),
-                                                    std::forward<Ts>(args)...);
-    case cudf::type_id::DURATION_NANOSECONDS:
-      return std::make_shared<arrow::DurationArray>(arrow::duration(arrow::TimeUnit::NANO),
-                                                    std::forward<Ts>(args)...);
-    default: CUDF_FAIL("Unsupported type_id conversion to arrow");
-  }
-}
 
 extern "C" {
 
@@ -198,27 +140,15 @@ JNIEXPORT jlong JNICALL Java_ai_rapids_cudf_ColumnVector_fromArrow(JNIEnv* env,
         break;
       default:
         // this handles the primitive types
-        arrow_array = to_arrow_array(n_type, j_col_length, data_buffer, null_buffer, j_null_count);
+        arrow_array = cudf::detail::to_arrow_array(
+          n_type, j_col_length, data_buffer, null_buffer, j_null_count);
     }
     auto name_and_type                                = arrow::field("col", arrow_array->type());
     std::vector<std::shared_ptr<arrow::Field>> fields = {name_and_type};
     std::shared_ptr<arrow::Schema> schema             = std::make_shared<arrow::Schema>(fields);
     auto arrow_table =
       arrow::Table::Make(schema, std::vector<std::shared_ptr<arrow::Array>>{arrow_array});
-
-    ArrowSchema sch;
-    if (!arrow::ExportSchema(*arrow_table->schema(), &sch).ok()) {
-      JNI_THROW_NEW(env, "java/lang/RuntimeException", "Unable to produce an ArrowSchema", 0)
-    }
-    auto batch = arrow_table->CombineChunksToBatch().ValueOrDie();
-    ArrowArray arr;
-    if (!arrow::ExportRecordBatch(*batch, &arr).ok()) {
-      JNI_THROW_NEW(env, "java/lang/RuntimeException", "Unable to produce an ArrowArray", 0)
-    }
-    auto retCols = cudf::from_arrow(&sch, &arr)->release();
-    arr.release(&arr);
-    sch.release(&sch);
-
+    auto retCols = cudf::from_arrow(*(arrow_table))->release();
     if (retCols.size() != 1) {
       JNI_THROW_NEW(env, "java/lang/IllegalArgumentException", "Must result in one column", 0);
     }