## Description
<!-- Provide a standalone description of changes in this PR. -->
<!-- Reference any issues closed by this PR with "closes #1234". -->
<!-- Note: The pull request title will be included in the CHANGELOG. -->
The intent of this PR is to fix #17075, but it provides no fixes as of yet. Due to the large changes that we have made around both arrow and pylibcudf since 24.08, getting back to a state where we can benchmark and profile cudf's legacy arrow interop against the current C data interface implementations is challenging. This PR makes the necessary changes to enable side-by-side benchmarking and profiling. The code will switch between the old and new implementation using the `USE_LEGACY_INTEROP` flag, like so:
```
(rapids) coder ➜ ~/cudf/arrow_slowdown $ ipython test.ipy 
to_arrow() time:
1.54 s ± 1.37 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
from_arrow() time:
587 ms ± 120 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
GPU -> CPU -> GPU Arrow roundtrip time:
2.03 s ± 8.61 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
to_pandas() time:
2.12 s ± 614 μs per loop (mean ± std. dev. of 2 runs, 2 loops each)
GPU -> CPU -> GPU Pandas roundtrip time:
3.26 s ± 47.6 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
(rapids) coder ➜ 
(rapids) coder ➜ ~/cudf/arrow_slowdown $ USE_LEGACY_INTEROP=1 ipython test.ipy 
to_arrow() time:
728 ms ± 1.23 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
from_arrow() time:
454 ms ± 7.49 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
GPU -> CPU -> GPU Arrow roundtrip time:
1.15 s ± 443 μs per loop (mean ± std. dev. of 2 runs, 2 loops each)
to_pandas() time:
2.11 s ± 395 μs per loop (mean ± std. dev. of 2 runs, 2 loops each)
GPU -> CPU -> GPU Pandas roundtrip time:
3.25 s ± 45.8 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
```

I'll move on to profiling from here, but this should also make it easier for someone else to pick up this benchmarking if necessary.

## Checklist
- [ ] I am familiar with the [Contributing Guidelines](https://github.com/rapidsai/cudf/blob/HEAD/CONTRIBUTING.md).
- [ ] New or existing tests cover these changes.
- [ ] The documentation is up to date with these changes.