@vyasr 2024-10-15
Using this branch and with the fix in https://github.com/rapidsai/cudf/pull/17081/commits/ff3e5df3531c7e9d260ac70e5b01ecdf8b713391, I now see the following:
```bash
(rapids) coder ➜ ~/cudf/arrow_slowdown $ USE_LEGACY_INTEROP=0 ipython interop-benchmark.ipy
to_arrow() time:
1.55 s ± 2.33 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
from_arrow() time:
1.19 s ± 170 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
GPU -> CPU -> GPU Arrow roundtrip time:
2.41 s ± 4.32 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
to_pandas() time:
4.08 s ± 19.8 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
GPU -> CPU -> GPU Pandas roundtrip time:
6.54 s ± 400 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
```
```bash
(rapids) coder ➜ ~/cudf/arrow_slowdown $ USE_LEGACY_INTEROP=1 ipython interop-benchmark.ipy
to_arrow() time:
1.4 s ± 2.24 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
from_arrow() time:
1.18 s ± 180 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
GPU -> CPU -> GPU Arrow roundtrip time:
2.26 s ± 1.18 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
to_pandas() time:
3.73 s ± 817 μs per loop (mean ± std. dev. of 2 runs, 2 loops each)
GPU -> CPU -> GPU Pandas roundtrip time:
6.24 s ± 390 ms per loop (mean ± std. dev. of 2 runs, 2 loops each)
```

These numbers vary enough from run to run to change which implementation is faster. The differences in these particular results are within the band of noise.